refactor(QuorumCustody): reorder ops, optimize functions

+++ contracts/evm/script/AddSigners.s.sol
@@ -23,7 +23,7 @@ contract AddSigners is Script {
         uint256 callerKey = vm.envUint("PRIVATE_KEY");
         address contractAddr = vm.envAddress("CONTRACT");
         address[] memory newSigners = vm.envAddress("NEW_SIGNERS", ",");
-        uint256 newQuorum = vm.envUint("NEW_QUORUM");
+        uint64 newQuorum = uint64(vm.envUint("NEW_QUORUM"));
         uint256 deadline = vm.envUint("DEADLINE");
 
         bytes[] memory sigs;
